feat(git-package): creation of a git package to centralize the codebase
The problem that this commit tries to solve is that I don't know how to emulate a lambda layer
locally using terraform.
So, by creating a git package, I can centralize do code that the lambda
files are going to import and we won't need to maintain a bunch of support structures.
I should
takke a deeper look into serverless environment using terraform alone