refactor(terraform): lambda layer generation attempt, still mannually ziping and publishing it
The lambda layer was the biggest challenge for me. The combination of terraform with Typescript
broke my project structure.
I know how to deploy the js files based on the ts, but was not easy to
wrap my head around it.

Specially since I missed the serverless offline ease of lambda invoke